Rename topic behavior file
