+++ dds/src/implementation/dds_impl/mod.rs
@@ -1,6 +1,10 @@
 pub mod any_data_reader_listener;
 pub mod any_data_writer_listener;
 pub mod any_topic_listener;
+pub mod behavior_builtin_data_reader_stateful;
+pub mod behavior_builtin_data_reader_stateless;
+pub mod behavior_user_defined_data_reader;
+pub mod behavior_user_defined_topic;
 pub mod dds_data_reader;
 pub mod dds_data_writer;
 pub mod dds_domain_participant;
@@ -9,16 +13,12 @@ pub mod dds_subscriber;
 pub mod dds_topic;
 pub mod iterators;
 pub mod message_receiver;
-pub mod behavior_builtin_data_reader_stateful;
-pub mod behavior_builtin_data_reader_stateless;
 pub mod node_builtin_subscriber;
 pub mod node_domain_participant;
-pub mod nodes;
-pub mod behavior_user_defined_data_reader;
 pub mod node_user_defined_data_writer;
 pub mod node_user_defined_publisher;
 pub mod node_user_defined_subscriber;
-pub mod node_user_defined_topic;
+pub mod nodes;
 pub mod participant_discovery;
 pub mod status_condition_impl;
 pub mod status_listener;
